server/requirements.txt
beautifulsoup4==4.6.0
dj-database-url==0.5.0
Django==1.11
django-cors-middleware==1.3.1
django-postgrespool==0.3.1
django-rq==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.7.7
djoser==1.1.5
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
html2text==2018.1.9
psycopg2-binary==2.7.4
PyJWT==1.6.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2018.3
raven==6.6.0
redis==2.10.6
requests==2.18.4
rq==0.10.0
